relu/fill: add f64-safe attrs and align Vulkan push constants
- add AttrValue::Double and parse DSL float literals as f64
- preserve f64 fill values via low/high push-constant packing
- rename ReLU negative_slope to alpha across DSL/docs/examples
- add f64-capable alpha/clamp handling in ReLU shaders and runtime
- refactor matmul shader layout and add f8/bf16 custom paths

+++ openinfer-dsl/src/parsers/op.rs
@@ -31,11 +31,11 @@ pub(crate) fn parse_op_attr_value(input: ParseStream) -> Result<OpAttrValue> {
 
     if input.peek(syn::LitFloat) {
         let lit: syn::LitFloat = input.parse()?;
-        let mut value: f32 = lit.base10_parse()?;
+        let mut value: f64 = lit.base10_parse()?;
         if negative {
             value = -value;
         }
-        return Ok(OpAttrValue::Float(value));
+        return Ok(OpAttrValue::Double(value));
     }
     if input.peek(syn::LitBool) {
         let lit: syn::LitBool = input.parse()?;
@@ -65,7 +65,7 @@ pub(crate) fn parse_op_attr_value(input: ParseStream) -> Result<OpAttrValue> {
             return Ok(OpAttrValue::Bool(false));
         }
         if name == "inf" {
-            return Ok(OpAttrValue::Float(f32::INFINITY));
+            return Ok(OpAttrValue::Double(f64::INFINITY));
         }
         return Ok(OpAttrValue::Var(ident));
     }
